Refactor hll_sktch and fix stack use after free bug
