+++ be/src/exec/aggregator.cpp
@@ -203,20 +203,6 @@ void AggregatorParams::init() {
 #define ALIGN_TO(size, align) ((size + align - 1) / align * align)
 #define PAD(size, align) (align - (size % align)) % align;
 
-class ThreadLocalStateAllocatorSetter {
-public:
-    ThreadLocalStateAllocatorSetter(Allocator* allocator)
-            : _agg_state_allocator_setter(allocator), _roaring_allocator_setter(allocator) {}
-    ~ThreadLocalStateAllocatorSetter() = default;
-
-private:
-    ThreadLocalAggregateStateAllocatorSetter _agg_state_allocator_setter;
-    ThreadLocalRoaringAllocatorSetter _roaring_allocator_setter;
-};
-
-#define SCOPED_THREAD_LOCAL_STATE_ALLOCATOR_SETTER(allocator) \
-    auto VARNAME_LINENUM(alloc_setter) = ThreadLocalStateAllocatorSetter(allocator)
-
 Aggregator::Aggregator(AggregatorParamsPtr params) : _params(std::move(params)) {
     _allocator = std::make_unique<CountingAllocatorWithHook>();
 }
@@ -395,59 +381,22 @@ Status Aggregator::prepare(RuntimeState* state, ObjectPool* pool, RuntimeProfile
         const TFunction& fn = desc.nodes[0].fn;
         const auto& agg_fn_type = _agg_fn_types[i];
         _is_merge_funcs[i] = aggregate_functions[i].nodes[0].agg_expr.is_merge_agg;
+
         // get function
-        if (fn.name.function_name == "count") {
-            bool is_input_nullable =
-                    !fn.arg_types.empty() && (has_outer_join_child || desc.nodes[0].has_nullable_child);
-            auto* func = get_aggregate_function("count", TYPE_BIGINT, TYPE_BIGINT, is_input_nullable);
-            _agg_functions[i] = func;
-        } else {
-            TypeDescriptor return_type = TypeDescriptor::from_thrift(fn.ret_type);
-            TypeDescriptor serde_type = TypeDescriptor::from_thrift(fn.aggregate_fn.intermediate_type);
-
-            TypeDescriptor arg_type = TypeDescriptor::from_thrift(fn.arg_types[0]);
-            // Because intersect_count have two input types.
-            // And intersect_count's first argument's type is alwasy Bitmap,
-            // so we use its second arguments type as input.
-            if (fn.name.function_name == "intersect_count") {
-                arg_type = TypeDescriptor::from_thrift(fn.arg_types[1]);
+        auto is_result_nullable_func = [&]() {
+            if (fn.name.function_name == "count") {
+                if (fn.arg_types.empty()) {
+                    return false;
+                }
+                if (has_outer_join_child || desc.nodes[0].has_nullable_child) {
+                    return true;
+                }
+                return false;
+            } else {
+                return agg_fn_type.use_nullable_fn(_use_intermediate_as_output());
             }
-
-            // Because max_by and min_by function have two input types,
-            // so we use its second arguments type as input.
-            if (fn.name.function_name == "max_by" || fn.name.function_name == "min_by") {
-                arg_type = TypeDescriptor::from_thrift(fn.arg_types[1]);
-            }
-
-            // Because windowfunnel have more two input types.
-            // functions registry use 2th args(datetime/date).
-            if (fn.name.function_name == "window_funnel") {
-                arg_type = TypeDescriptor::from_thrift(fn.arg_types[1]);
-            }
-
-            // hack for accepting various arguments
-            if (fn.name.function_name == "exchange_bytes" || fn.name.function_name == "exchange_speed") {
-                arg_type = TypeDescriptor(TYPE_BIGINT);
-            }
-
-            if (fn.name.function_name == "array_union_agg" || fn.name.function_name == "array_unique_agg") {
-                // for array_union_agg use inner type as signature
-                arg_type = arg_type.children[0];
-            }
-
-            const bool use_nullable_fn = agg_fn_type.use_nullable_fn(_use_intermediate_as_output());
-            auto* func = get_aggregate_function(fn.name.function_name, arg_type.type, return_type.type, use_nullable_fn,
-                                                fn.binary_type, state->func_version());
-            if (func == nullptr) {
-                return Status::InternalError(strings::Substitute(
-                        "Invalid agg function plan: $0 with (arg type $1, serde type $2, result type $3, nullable $4)",
-                        fn.name.function_name, type_to_string(arg_type.type), type_to_string(serde_type.type),
-                        type_to_string(return_type.type), use_nullable_fn ? "true" : "false"));
-            }
-            VLOG_ROW << "get agg function " << func->get_name() << " serde_type " << serde_type << " return_type "
-                     << return_type;
-            _agg_functions[i] = func;
-        }
+        };
+        RETURN_IF_ERROR(_create_aggregate_function(state, fn, is_result_nullable_func(), &_agg_functions[i]));
 
         int node_idx = 0;
         for (int j = 0; j < desc.nodes[0].num_children; ++j) {
@@ -527,6 +476,42 @@ Status Aggregator::prepare(RuntimeState* state, ObjectPool* pool, RuntimeProfile
         RETURN_IF_ERROR(spiller()->prepare(state));
     }
 
+    return Status::OK();
+}
+
+Status Aggregator::_create_aggregate_function(starrocks::RuntimeState* state, const TFunction& fn,
+                                              bool is_result_nullable, const AggregateFunction** ret) {
+    std::vector<TypeDescriptor> arg_types;
+    for (auto& type : fn.arg_types) {
+        arg_types.push_back(TypeDescriptor::from_thrift(type));
+    }
+
+    // check whether it's _merge/_union combinator if it contains agg state type
+    auto& func_name = fn.name.function_name;
+    // get function
+    if (func_name == "count") {
+        auto* func = get_aggregate_function("count", TYPE_BIGINT, TYPE_BIGINT, is_result_nullable);
+        if (func == nullptr) {
+            return Status::InternalError(strings::Substitute("Invalid agg function plan: $0 ", func_name));
+        }
+        *ret = func;
+    } else {
+        TypeDescriptor return_type = TypeDescriptor::from_thrift(fn.ret_type);
+        TypeDescriptor serde_type = TypeDescriptor::from_thrift(fn.aggregate_fn.intermediate_type);
+        DCHECK_LE(1, fn.arg_types.size());
+        TypeDescriptor arg_type = arg_types[0];
+        auto* func = get_aggregate_function(func_name, return_type, arg_types, is_result_nullable, fn.binary_type,
+                                            state->func_version());
+        if (func == nullptr) {
+            return Status::InternalError(strings::Substitute(
+                    "Invalid agg function plan: $0 with (arg type $1, serde type $2, result type $3, nullable $4)",
+                    func_name, type_to_string(arg_type.type), type_to_string(serde_type.type),
+                    type_to_string(return_type.type), is_result_nullable ? "true" : "false"));
+        }
+        *ret = func;
+        VLOG_ROW << "get agg function " << func->get_name() << " serde_type " << serde_type << " return_type "
+                 << return_type;
+    }
     return Status::OK();
 }
