+++ be/src/types/CMakeLists.txt
@@ -19,6 +19,7 @@ add_library(Types STATIC
     bitmap_value.cpp
     date_value.cpp
     hll.cpp
+    hll_sketch.cpp
     logical_type.cpp
     map_type_info.cpp
     struct_type_info.cpp
